m_bus: widen pl_offset to avoid uint8_t wraparound
pl_offset is pl_base plus the header size. pl_base grows through the AFL
sub-header recursion by 2 + afl_len per level, and afl_len is a byte off
the wire. One AFL on a full length telegram pushes the total past 255.
A uint8_t keeps only the low 8 bits, so parse_payload starts its DIF/VIF
walk from the wrapped value and reads link layer header bytes as data
records.

Same shape as the parse_payload off fix in 11895fe. The bounds check was
fine. The type holding the value it checks was too narrow.

+++ src/devices/m_bus.c
@@ -148,7 +148,7 @@ typedef struct {
     uint8_t     AC;         // Access number
     uint8_t     ST;
     uint16_t    CW;         // Configuration word
-    uint8_t     pl_offset;  // Payload offset
+    unsigned    pl_offset;  // Payload offset
     /* Extended Link Layer (EN 13757-4), when the frame has one: this
        wraps the actual Transport/Application Layer CI above, which is
        parsed from just after it once found not to be encrypted. */
